Quit button on Client and Server both work.
Takes user back to main screen and a new game can be started.

Device Bluetooth name Adds "Social Play Server" to the front of its name so that if more than 1 is available they can be differentiated. After the game is complete, it name returns to what it once was.

Fixed issue where client would connect to device even if it was not available. No the device on the list will light up, but it will notify that it cannot connect and remain on client screen.

Ensured that server has no background threads running that client can possibly connect to within the app.

TODO: When a player goes from a in one game to the server in the next the previous server is already bonded with a none Social Play name so he must start and restart the server fragment to connect. Not sure how to fix this other than unbonding all devices at the end of the game, which is possible but brings up the issue of all the dialogs when the game is played.
TODO: Make it more visually appealing.
TODO: Error Handling.